rename ibc-link to link
