+++ link/internal/network/tools.go
